Add AdminHoldingsPage
